Front end: Se agregaron buscadores, Ajustes en tablas generales, Se agrego vista de cliente para las cotizaciones de impuestos. Se cambiaron detalles en diseño.
Back-end:
Se agregaron los store procedure de revisionalmacen, se agrego el funcionamiento del CRUD desde la Api

+++ Back-end/Samusa-Back/Samusa-Back/Models/RevisionAlmacen.cs
@@ -15,7 +15,7 @@ public partial class RevisionAlmacen
 
     public string Color { get; set; } = null!;
 
-    public decimal CostoVehiculo { get; set; }
+    public int CostoVehiculo { get; set; }
 
     public int AnioVehiculo { get; set; }
